+++ pom.xml
@@ -9,10 +9,10 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>co.inventrosoft</groupId>
-    <artifactId>jdbc-task</artifactId>
+    <artifactId>hibernate-task</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>jdbc-task</name>
-    <description>Demo project for JDBC</description>
+    <name>hibernate-task</name>
+    <description>Demo project for Hibernate</description>
     <properties>
         <java.version>16</java.version>
 
@@ -23,7 +23,14 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-devtools</artifactId>
