+++ src/main/java/co/inventrosoft/hibernate_task/configuration/AppConfiguration.java
@@ -1,0 +1,14 @@
+package co.inventrosoft.hibernate_task.configuration;
+
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfiguration {
+    @Bean
+    public ObjectMapper mapper() {
+        return new ObjectMapper();
+    }
+}
