+++ src/main/java/co/inventrosoft/hibernate_task/configuration/JdbcConfig.java
@@ -1,0 +1,21 @@
+package co.inventrosoft.hibernate_task.configuration;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "spring.datasource")
+public class JdbcConfig {
+    @Value("${spring.datasource.driver-class-name}")
+    private String driverClassName;
+    @Value("${spring.datasource.url}")
+    private String url;
+    @Value("${spring.datasource.username}")
+    private String username;
+    @Value("${spring.datasource.password}")
+    private String password;
+}
