+++ src/main/java/co/inventrosoft/hibernate_task/model/Tournament.java
@@ -1,0 +1,23 @@
+package co.inventrosoft.hibernate_task.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString(of = "id")
+@Table(name = "tournaments")
+public class Tournament {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(unique = true, nullable = false)
+    int id;
+
+}
