+++ src/main/java/co/inventrosoft/hibernate_task/repository/TeamRepository.java
@@ -1,0 +1,12 @@
+package co.inventrosoft.hibernate_task.repository;
+
+import co.inventrosoft.hibernate_task.model.Team;
+
+import java.util.List;
+
+public interface TeamRepository {
+    void save(List<Team> teams);
+    void add(Team team);
+    void update(Team team);
+    boolean isExist(String teamName);
+}
