+++ src/main/java/co/inventrosoft/hibernate_task/repository/TournamentRepository.java
@@ -1,0 +1,6 @@
+package co.inventrosoft.hibernate_task.repository;
+
+
+public interface TournamentRepository {
+    int create();
+}
