+++ src/main/java/co/inventrosoft/hibernate_task/repository/hibernate/TeamRepository.java
@@ -1,0 +1,10 @@
+package co.inventrosoft.hibernate_task.repository.hibernate;
+
+import co.inventrosoft.hibernate_task.model.Team;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TeamRepository extends JpaRepository<Team, Integer> {
+    boolean existsByName(String name);
+}
