+++ src/main/java/co/inventrosoft/hibernate_task/repository/hibernate/TournamentRepository.java
@@ -1,0 +1,12 @@
+package co.inventrosoft.hibernate_task.repository.hibernate;
+
+
+import co.inventrosoft.hibernate_task.model.Tournament;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface TournamentRepository extends JpaRepository<Tournament, Integer> {
+
+}
