--- src/main/java/co/inventrosoft/springboottask/configuration/AppConfiguration.java
@@ -1,14 +1,0 @@
-package co.inventrosoft.springboottask.configuration;
-
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AppConfiguration {
-    @Bean
-    public ObjectMapper mapper() {
-        return new ObjectMapper();
-    }
-}
