--- src/main/java/co/inventrosoft/springboottask/repository/TournamentRepository.java
@@ -1,10 +1,0 @@
-package co.inventrosoft.springboottask.repository;
-
-
-import co.inventrosoft.springboottask.model.Match;
-
-import java.util.List;
-
-public interface TournamentRepository {
-    int create();
-}
