+++ src/main/resources/application.properties
@@ -1,8 +1,14 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/spring-jdbc-task
 spring.datasource.username=root
 spring.datasource.password=123
-
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-spring.sql.init.mode=always
-spring.sql.init.schema-locations=classpath*:initDB.sql
+#spring.sql.init.mode=always
+#spring.sql.init.schema-locations=classpath*:initDB.sql
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.open-in-view=false
+spring.jpa.hibernate.ddl-auto=update
+spring.data.jpa.repositories.bootstrap-mode=lazy
+spring.main.lazy-initialization=true
+spring.jmx.enabled=false
